+++ src/mcmc/repeat_sampler.jl
@@ -24,16 +24,12 @@ struct RepeatSampler{S<:AbstractMCMC.AbstractSampler} <: AbstractMCMC.AbstractSa
     end
 end
 
-function RepeatSampler(alg::InferenceAlgorithm, num_repeat::Int)
-    return RepeatSampler(Sampler(alg), num_repeat)
-end
-
-getADType(spl::RepeatSampler) = getADType(spl.sampler)
-DynamicPPL.default_chain_type(sampler::RepeatSampler) = default_chain_type(sampler.sampler)
-# TODO(mhauru) Remove the below once DynamicPPL has removed all its Selector stuff.
-DynamicPPL.inspace(vn::VarName, spl::RepeatSampler) = inspace(vn, spl.sampler)
-
-function setparams_varinfo!!(model::DynamicPPL.Model, sampler::RepeatSampler, state, params)
+function setparams_varinfo!!(
+    model::DynamicPPL.Model,
+    sampler::RepeatSampler,
+    state,
+    params::DynamicPPL.AbstractVarInfo,
+)
     return setparams_varinfo!!(model, sampler.sampler, state, params)
 end
 
@@ -42,6 +38,14 @@ function AbstractMCMC.step(
     model::AbstractMCMC.AbstractModel,
     sampler::RepeatSampler;
     kwargs...,
+)
+    return AbstractMCMC.step(rng, model, sampler.sampler; kwargs...)
+end
+# The following method needed for method ambiguity resolution.
+# TODO(penelopeysm): Remove this method once the default `AbstractMCMC.step(rng,
+# ::DynamicPPL.Model, ::AbstractSampler)` method in `src/mcmc/abstractmcmc.jl` is removed.
+function AbstractMCMC.step(
+    rng::Random.AbstractRNG, model::DynamicPPL.Model, sampler::RepeatSampler; kwargs...
 )
     return AbstractMCMC.step(rng, model, sampler.sampler; kwargs...)
 end
@@ -85,4 +89,63 @@ function AbstractMCMC.step_warmup(
         )
     end
     return transition, state
+end
+
+# Need some extra leg work to make RepeatSampler work seamlessly with DynamicPPL models +
+# samplers, instead of generic AbstractMCMC samplers.
+
+function Turing.Inference.init_strategy(spl::RepeatSampler)
+    return Turing.Inference.init_strategy(spl.sampler)
+end
+
+function AbstractMCMC.sample(
+    rng::AbstractRNG,
+    model::DynamicPPL.Model,
+    sampler::RepeatSampler,
+    N::Integer;
+    check_model=true,
+    initial_params=Turing.Inference.init_strategy(sampler),
+    chain_type=DEFAULT_CHAIN_TYPE,
+    progress=PROGRESS[],
+    kwargs...,
+)
+    check_model && _check_model(model, sampler)
+    return AbstractMCMC.mcmcsample(
+        rng,
+        model,
+        sampler,
+        N;
+        initial_params=_convert_initial_params(initial_params),
+        chain_type=chain_type,
+        progress=progress,
+        kwargs...,
+    )
+end
+
+function AbstractMCMC.sample(
+    rng::AbstractRNG,
+    model::DynamicPPL.Model,
+    sampler::RepeatSampler,
+    ensemble::AbstractMCMC.AbstractMCMCEnsemble,
+    N::Integer,
+    n_chains::Integer;
+    check_model=true,
+    initial_params=fill(Turing.Inference.init_strategy(sampler), n_chains),
+    chain_type=DEFAULT_CHAIN_TYPE,
+    progress=PROGRESS[],
+    kwargs...,
+)
+    check_model && _check_model(model, sampler)
+    return AbstractMCMC.mcmcsample(
+        rng,
+        model,
+        sampler,
+        ensemble,
+        N,
+        n_chains;
+        initial_params=map(_convert_initial_params, initial_params),
+        chain_type=chain_type,
+        progress=progress,
+        kwargs...,
+    )
 end