+++ test/mcmc/mh.jl
@@ -4,7 +4,6 @@ using AdvancedMH: AdvancedMH
 using Distributions:
     Bernoulli, Dirichlet, Exponential, InverseGamma, LogNormal, MvNormal, Normal, sample
 using DynamicPPL: DynamicPPL
-using DynamicPPL: Sampler
 using LinearAlgebra: I
 using Random: Random
 using StableRNGs: StableRNG
@@ -49,7 +48,7 @@ GKernel(var) = (x) -> Normal(x, sqrt.(var))
         # Set the initial parameters, because if we get unlucky with the initial state,
         # these chains are too short to converge to reasonable numbers.
         discard_initial = 1_000
-        initial_params = [1.0, 1.0]
+        initial_params = InitFromParams((s=1.0, m=1.0))
 
         @testset "gdemo_default" begin
             alg = MH()
@@ -72,7 +71,7 @@ GKernel(var) = (x) -> Normal(x, sqrt.(var))
             chain = sample(
                 StableRNG(seed), gdemo_default, alg, 10_000; discard_initial, initial_params
             )
-            check_gdemo(chain; atol=0.1)
+            check_gdemo(chain; atol=0.15)
         end
 
         @testset "MoGtest_default with Gibbs" begin
@@ -81,13 +80,16 @@ GKernel(var) = (x) -> Normal(x, sqrt.(var))
                 @varname(mu1) => MH((:mu1, GKernel(1))),
                 @varname(mu2) => MH((:mu2, GKernel(1))),
             )
+            initial_params = InitFromParams((
+                mu1=1.0, mu2=1.0, z1=0.0, z2=0.0, z3=1.0, z4=1.0
+            ))
             chain = sample(
                 StableRNG(seed),
                 MoGtest_default,
                 gibbs,
                 500;
                 discard_initial=100,
-                initial_params=[1.0, 1.0, 0.0, 0.0, 1.0, 4.0],
+                initial_params=initial_params,
             )
             check_MoGtest_default(chain; atol=0.2)
         end
@@ -113,7 +115,7 @@ GKernel(var) = (x) -> Normal(x, sqrt.(var))
         end
 
         model = M(zeros(2), I, 1)
-        sampler = Inference.Sampler(MH())
+        sampler = MH()
 
         dt, vt = Inference.dist_val_tuple(sampler, DynamicPPL.VarInfo(model))
 
@@ -184,7 +186,7 @@ GKernel(var) = (x) -> Normal(x, sqrt.(var))
         # Test that the small variance version is actually smaller.
         variance_small = var(diff(Array(chn_small["μ[1]"]); dims=1))
         variance_big = var(diff(Array(chn_big["μ[1]"]); dims=1))
-        @test variance_small < variance_big / 1_000.0
+        @test variance_small < variance_big / 100.0
     end
 
     @testset "vector of multivariate distributions" begin
@@ -228,38 +230,34 @@ GKernel(var) = (x) -> Normal(x, sqrt.(var))
         # Don't link when no proposals are given since we're using priors
         # as proposals.
         vi = deepcopy(vi_base)
-        alg = MH()
-        spl = DynamicPPL.Sampler(alg)
-        vi = Turing.Inference.maybe_link!!(vi, spl, alg.proposals, gdemo_default)
-        @test !DynamicPPL.islinked(vi)
+        spl = MH()
+        vi = Turing.Inference.maybe_link!!(vi, spl, spl.proposals, gdemo_default)
+        @test !DynamicPPL.is_transformed(vi)
 
         # Link if proposal is `AdvancedHM.RandomWalkProposal`
         vi = deepcopy(vi_base)
         d = length(vi_base[:])
-        alg = MH(AdvancedMH.RandomWalkProposal(MvNormal(zeros(d), I)))
-        spl = DynamicPPL.Sampler(alg)
-        vi = Turing.Inference.maybe_link!!(vi, spl, alg.proposals, gdemo_default)
-        @test DynamicPPL.islinked(vi)
+        spl = MH(AdvancedMH.RandomWalkProposal(MvNormal(zeros(d), I)))
+        vi = Turing.Inference.maybe_link!!(vi, spl, spl.proposals, gdemo_default)
+        @test DynamicPPL.is_transformed(vi)
 
         # Link if ALL proposals are `AdvancedHM.RandomWalkProposal`.
         vi = deepcopy(vi_base)
-        alg = MH(:s => AdvancedMH.RandomWalkProposal(Normal()))
-        spl = DynamicPPL.Sampler(alg)
-        vi = Turing.Inference.maybe_link!!(vi, spl, alg.proposals, gdemo_default)
-        @test DynamicPPL.islinked(vi)
+        spl = MH(:s => AdvancedMH.RandomWalkProposal(Normal()))
+        vi = Turing.Inference.maybe_link!!(vi, spl, spl.proposals, gdemo_default)
+        @test DynamicPPL.is_transformed(vi)
 
         # Don't link if at least one proposal is NOT `RandomWalkProposal`.
         # TODO: make it so that only those that are using `RandomWalkProposal`
         # are linked! I.e. resolve https://github.com/TuringLang/Turing.jl/issues/1583.
         # https://github.com/TuringLang/Turing.jl/pull/1582#issuecomment-817148192
         vi = deepcopy(vi_base)
-        alg = MH(
+        spl = MH(
             :m => AdvancedMH.StaticProposal(Normal()),
             :s => AdvancedMH.RandomWalkProposal(Normal()),
         )
-        spl = DynamicPPL.Sampler(alg)
-        vi = Turing.Inference.maybe_link!!(vi, spl, alg.proposals, gdemo_default)
-        @test !DynamicPPL.islinked(vi)
+        vi = Turing.Inference.maybe_link!!(vi, spl, spl.proposals, gdemo_default)
+        @test !DynamicPPL.is_transformed(vi)
     end
 
     @testset "`filldist` proposal (issue #2180)" begin